Fix context not updated while loging-in
